Rename precedence.js to more generic operators.js
