+++ lib/expression/node/AssignmentNode.js
@@ -4,7 +4,7 @@ var Node = require('./Node'),
     ArrayNode = require('./ArrayNode'),
 
     keywords = require('../keywords'),
-    precedence = require('../precedence'),
+    operators = require('../operators'),
 
     latex = require('../../util/latex'),
     isString = require('../../util/string').isString;
@@ -99,7 +99,7 @@ AssignmentNode.prototype.toTex = function() {
  * @return {Number}
  */
 AssignmentNode.prototype.getPrecedence = function () {
-  return precedence.getPrecedence(this);
+  return operators.getPrecedence(this);
 };
 
 module.exports = AssignmentNode;
