+++ lib/expression/node/ConditionalNode.js
@@ -6,7 +6,7 @@ var BigNumber = require('../../type/BigNumber');
 var Complex = require('../../type/Complex');
 var Unit = require('../../type/Unit');
 var util = require('../../util');
-var precedence = require('../precedence');
+var operators = require('../operators');
 
 var isString = util.string.isString;
 var isNumber = util.number.isNumber;
@@ -153,7 +153,7 @@ ConditionalNode.prototype.toTex = function() {
  * @return {Number}
  */
 ConditionalNode.prototype.getPrecedence = function () {
-  return precedence.getPrecedence(this);
+  return operators.getPrecedence(this);
 };
 
 module.exports = ConditionalNode;
