+++ lib/expression/node/OperatorNode.js
@@ -5,7 +5,7 @@ var Node = require('./Node'),
     SymbolNode = require('./SymbolNode'),
     FunctionNode = require('./FunctionNode'),
     latex = require('../../util/latex'),
-    precedence = require('../precedence');
+    operators = require('../operators');
 
 /**
  * @constructor OperatorNode
@@ -231,7 +231,7 @@ OperatorNode.prototype.getIdentifier = function() {
  * @return {Number}
  */
 OperatorNode.prototype.getPrecedence = function () {
-  return precedence.getPrecedence(this);
+  return operators.getPrecedence(this);
 };
 
 module.exports = OperatorNode;
