+++ lib/expression/node/RangeNode.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var Node = require('./Node');
-var precedence = require('../precedence');
+var operators = require('../operators');
 
 var isNode = Node.isNode;
 
@@ -117,7 +117,7 @@ RangeNode.prototype.toTex = function() {
  * @return {Number}
  */
 RangeNode.prototype.getPrecedence = function () {
-  return precedence.getPrecedence(this);
+  return operators.getPrecedence(this);
 };
 
 module.exports = RangeNode;
